requirements.txt
Bottleneck @ file:///C:/ci/bottleneck_1648029103672/work
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
django-crispy-forms==1.9.2
idna==3.3
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
numexpr @ file:///C:/ci/numexpr_1640704408721/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1652802206636/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/ci/pandas_1650373491186/work
pycparser==2.21
pyodbc @ file:///C:/ci/pyodbc_1647426060948/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==0.20.0
pytz==2021.3
requests==2.27.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
tableauhyperapi==0.0.14751
tableauserverclient==0.18.0
urllib3==1.26.9
wincertstore==0.2
